Refactor payment state tracking by introducing WooPosPaymentStateAnalyticsTracker
